requirements.txt
Django==2.2.24
django-crispy-forms==1.13.0
djangorestframework==3.12.4
mysqlclient==2.0.3
Pillow==8.4.0
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.2
